remove golangci-lint dependencies
Currently we import the package
"github.com/golangci/golangci-lint/cmd/golangci-lint" in the file
tools.go (which is excluded from the build using a build tag), in order
to be able to `go run` this tool from the Makefile.

However, this package pulls in a huge number of transitive dependencies,
and worse, anyone who adds this SDK package to their project will also
inherit all of these transitive dependencies.

In order to make the SDK more light-weight, let's remove this import.

Instead, update the Makefile to `go run` the specific version of
golangci-lint defined in the .github/workflow/lint.yaml configuration
(so that running `make lint` locally should run the same version of
golangci-lint that the GitHub Action will run).